requirements.txt
yfinance
anthropic
requests
python-dotenv
pyyaml
pandas
